# docs/api/openapi.yaml
openapi: '3.1.1'
info:
  title: A365 Gateway API
  description: 'Connect independently hosted agents to Microsoft Entra, Agent 365, Purview and Prompt Shields. Administration uses an Entra access token; agent ingress uses a gateway-issued secret. Both use Authorization: Bearer, but are not interchangeable. Evaluate → call your model only on allow → submit the interaction. HTTP 202 means queued, never confirmed delivery. All timestamps are UTC. Interactive requests act on the live deployment.'
  version: v1
servers:
  - url: http://127.0.0.1:5080/
paths:
...
  '/api/v1/agents/{agentId}/purview-policies':
    get:
      tags:
        - AgentPolicyAssignments
      summary: Read individual-agent policy assignments
      description: "Reports assignment states, binding freshness and observed allow/block timestamps. Assigned, synchronized and verified enforcement are distinct. A block does not identify which applicable policy caused it.\n\nAuthorization policy: AdministratorOnly."
      operationId: AgentPolicyAssignments.List
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AgentPolicyAssignmentsResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPolicyAssignmentsResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AgentPolicyAssignmentsResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}/purview-policies/review':
    post:
      tags:
        - AgentPolicyAssignments
      summary: Review an individual-agent assignment
      description: "Binds a compatible policy revision, administrator and exact agent identity to a short-lived review. No policy write occurs until confirmation. Use policyId and revision from the current catalog.\n\nAuthorization policy: AdministratorOnly."
      operationId: AgentPolicyAssignments.Review
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Selection'
          text/json:
            schema:
              $ref: '#/components/schemas/Selection'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Selection'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentReviewResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentReviewResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentReviewResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
  '/api/v1/agents/{agentId}/purview-policies/{operationId}/confirm':
    post:
      tags:
        - AgentPolicyAssignments
      summary: Confirm a reviewed policy assignment
      description: "Queues the reviewed assignment for this individual identity. Preserves unrelated targets, exclusions, rules and mode. Returns 202 for new work or 200 for an already confirmed operation. Poll assignments for confirmation; requests fail closed while required protection is unavailable.\n\nAuthorization policy: AdministratorOnly."
      operationId: AgentPolicyAssignments.Confirm
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentConfirmationResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentConfirmationResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentConfirmationResponse'
        '202':
          description: Accepted
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentConfirmationResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentConfirmationResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignmentConfirmationResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: 'Missing, invalid or expired credential. Use the authentication scheme documented for this operation.'
        '403':
          description: 'Authenticated caller is not permitted, or protection blocked the request. Inspect the response errorCode and correlationId.'
        '429':
          description: Gateway rate limit exceeded. Respect Retry-After when supplied.
      security:
        - EntraBearer: [ ]
...
components:
  schemas:
...
    AgentPolicyAssignmentResponse:
      required:
        - operationId
        - policyId
        - policyName
        - status
        - failureCode
        - confirmedAtUtc
        - expiresAtUtc
        - assignedAtUtc
      type: object
      properties:
        operationId:
          type: string
          format: uuid
        policyId:
          type: string
          format: uuid
        policyName:
          type: string
        status:
          type: string
        failureCode:
          type:
            - 'null'
            - string
        confirmedAtUtc:
          type:
            - 'null'
            - string
          format: date-time
        expiresAtUtc:
          type: string
          format: date-time
        assignedAtUtc:
          type:
            - 'null'
            - string
          format: date-time
    AgentPolicyAssignmentsResponse:
      required:
        - agentId
        - agentIdentityId
        - bindingCurrent
        - allowObservedAtUtc
        - blockObservedAtUtc
        - items
      type: object
      properties:
        agentId:
          type: string
          format: uuid
        agentIdentityId:
          type:
            - 'null'
            - string
        bindingCurrent:
          type: boolean
        allowObservedAtUtc:
          type:
            - 'null'
            - string
          format: date-time
        blockObservedAtUtc:
          type:
            - 'null'
            - string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgentPolicyAssignmentResponse'
...
    PolicyAssignmentConfirmationResponse:
      required:
        - operationId
        - status
      type: object
      properties:
        operationId:
          type: string
          format: uuid
        status:
          type: string
    PolicyAssignmentReviewResponse:
      required:
        - operationId
        - agentId
        - agentIdentityId
        - policyId
        - policyName
        - expiresAtUtc
        - effect
      type: object
      properties:
        operationId:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        agentIdentityId:
          type: string
          format: uuid
        policyId:
          type: string
          format: uuid
        policyName:
          type: string
        expiresAtUtc:
          type: string
          format: date-time
        effect:
          type: string
...
    ProblemDetails:
      type: object
      properties:
        type:
          type:
            - 'null'
            - string
        title:
          type:
            - 'null'
            - string
        status:
          pattern: '^-?(?:0|[1-9]\d*)$'
          type:
            - 'null'
            - integer
            - string
          format: int32
        detail:
          type:
            - 'null'
            - string
        instance:
          type:
            - 'null'
            - string
...
    Selection:
      required:
        - policyId
        - revision
      type: object
      properties:
        policyId:
          type: string
          format: uuid
        revision:
          type: string
...
  securitySchemes:
    EntraBearer:
      type: http
      description: Entra access token for the Gateway API audience with the endpoint's required gateway role. Do not paste a gateway agent key here.
      scheme: bearer
      bearerFormat: JWT
    AgentKey:
      type: http
      description: One-time gateway-issued agent key (a365gw_v1_...). Scoped to its registered agent. Do not use an Entra token for agent ingress.
      scheme: bearer
...